requirements.txt
# Salesforce API
simple-salesforce>=1.12.0

# スクレイピング
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# データ処理
pandas>=2.0.0
numpy>=1.24.0

# 設定管理
python-dotenv>=1.0.0
pyyaml>=6.0.0

# ロギング
loguru>=0.7.0

# スケジューリング
schedule>=1.2.0

# 型ヒント
pydantic>=2.5.0

# テスト
pytest>=7.4.0
pytest-cov>=4.1.0

# 開発ツール
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
